refactor: idのValueObjectの基底クラスを作成
