+++ src/domain-models/entities/node.ts
@@ -1,8 +1,8 @@
 import type { NodeRole } from "../value-objects/node-role";
 import type { NodeRenderingProperty } from "../value-objects/node-rendering-property";
-import type { NodeId } from "../value-objects/node-id";
-import type { ProjectId } from "../value-objects/project-id";
-import type { EdgeId } from "../value-objects/edge-id";
+import type { NodeId } from "../value-objects/id-value-objects/node-id";
+import type { ProjectId } from "../value-objects/id-value-objects/project-id";
+import type { EdgeId } from "../value-objects/id-value-objects/edge-id";
 
 export class Node {
   public readonly id: NodeId;
