+++ src/domain-models/entities/nodes/ai-node.ts
@@ -1,9 +1,9 @@
 import type { AINodeProperty } from "../../value-objects/ai-node-property";
 import type { AINodeState } from "../../value-objects/ai-node-state";
 import type { NodeRenderingProperty } from "../../value-objects/node-rendering-property";
-import type { ProjectId } from "../../value-objects/project-id";
+import type { ProjectId } from "../../value-objects/id-value-objects/project-id";
 import { Node } from "../node";
-import type { NodeId } from "../../value-objects/node-id";
+import type { NodeId } from "../../value-objects/id-value-objects/node-id";
 
 export class AINode extends Node {
   private _property: AINodeProperty;
