+++ src/domain-models/entities/nodes/system-node.ts
@@ -1,7 +1,7 @@
-import type { NodeId } from "@/domain-models/value-objects/node-id";
+import type { NodeId } from "@/domain-models/value-objects/id-value-objects/node-id";
 import { Node } from "../node";
 import type { NodeRenderingProperty } from "@/domain-models/value-objects/node-rendering-property";
-import type { ProjectId } from "@/domain-models/value-objects/project-id";
+import type { ProjectId } from "@/domain-models/value-objects/id-value-objects/project-id";
 
 export class SystemNode extends Node {
   constructor(
