+++ src/domain-models/entities/project.ts
@@ -1,7 +1,7 @@
-import type { ProjectId } from "../value-objects/project-id";
-import type { NodeId } from "../value-objects/node-id";
-import type { EdgeId } from "../value-objects/edge-id";
-import type { ContextId } from "../value-objects/context-id";
+import type { ProjectId } from "../value-objects/id-value-objects/project-id";
+import type { NodeId } from "../value-objects/id-value-objects/node-id";
+import type { EdgeId } from "../value-objects/id-value-objects/edge-id";
+import type { ContextId } from "../value-objects/id-value-objects/context-id";
 
 export class Project {
   public readonly id: ProjectId;
