+++ src/domain-models/value-objects/command-units/append-context-edge.ts
@@ -1,6 +1,6 @@
 import type { CommandUnit } from "../command-unit";
-import type { ContextId } from "../context-id";
-import type { EdgeId } from "../edge-id";
+import type { ContextId } from "../id-value-objects/context-id";
+import type { EdgeId } from "../id-value-objects/edge-id";
 
 export class AppendContextEdge implements CommandUnit {
   public readonly type = "appendContextEdge";
