+++ src/domain-models/value-objects/command-units/complete-ai-node-calculation.ts
@@ -1,5 +1,5 @@
 import type { CommandUnit } from "../command-unit";
-import type { NodeId } from "../node-id";
+import type { NodeId } from "../id-value-objects/node-id";
 
 export class CompleteAINodeCalculation implements CommandUnit {
   public readonly type = "completeAINodeCalculation";
