+++ src/domain-models/value-objects/command-units/create-context.ts
@@ -1,7 +1,7 @@
 import type { CommandUnit } from "../command-unit";
-import type { ContextId } from "../context-id";
+import type { ContextId } from "../id-value-objects/context-id";
 import type { NodeRenderingProperty } from "../node-rendering-property";
-import type { ProjectId } from "../project-id";
+import type { ProjectId } from "../id-value-objects/project-id";
 
 export class CreateContext implements CommandUnit {
   public readonly type = "createContext";
