+++ src/domain-models/value-objects/command-units/update-ai-node-property.ts
@@ -1,6 +1,6 @@
 import type { CommandUnit } from "../command-unit";
 import type { AINodeProperty } from "../ai-node-property";
-import type { NodeId } from "../node-id";
+import type { NodeId } from "../id-value-objects/node-id";
 
 export class UpdateAINodeProperty implements CommandUnit {
   public readonly type = "updateAINodeProperty";
