+++ src/domain-models/value-objects/id-value-objects/context-id.ts
@@ -1,0 +1,5 @@
+import { IDValueObject } from "../id-value-object";
+
+export class ContextId extends IDValueObject {
+  protected readonly _brand = "ContextId";
+}
