+++ src/domain-models/value-objects/id-value-objects/edge-id.ts
@@ -1,0 +1,5 @@
+import { IDValueObject } from "../id-value-object";
+
+export class EdgeId extends IDValueObject {
+  protected readonly _brand = "EdgeId";
+}
