+++ src/domain-models/value-objects/id-value-objects/node-id.ts
@@ -1,0 +1,5 @@
+import { IDValueObject } from "../id-value-object";
+
+export class NodeId extends IDValueObject {
+  protected readonly _brand = "NodeId";
+}
