+++ src/domain-models/value-objects/id-value-objects/project-id.ts
@@ -1,0 +1,5 @@
+import { IDValueObject } from "../id-value-object";
+
+export class ProjectId extends IDValueObject {
+  protected readonly _brand = "ProjectId";
+}
